+++ cdp-agentkit-core/cdp_agentkit_core/actions/social/twitter/__init__.py
@@ -1,6 +1,8 @@
 from cdp_agentkit_core.actions.social.twitter.account_details import AccountDetailsAction
+from cdp_agentkit_core.actions.social.twitter.account_mentions import AccountMentionsAction
 from cdp_agentkit_core.actions.social.twitter.action import TwitterAction
 from cdp_agentkit_core.actions.social.twitter.post_tweet import PostTweetAction
+from cdp_agentkit_core.actions.social.twitter.post_tweet_reply import PostTweetReplyAction
 
 
 def get_all_twitter_actions() -> list[type[TwitterAction]]:
@@ -16,6 +18,8 @@ TWITTER_ACTIONS = get_all_twitter_actions()
 __all__ = [
     "TwitterAction",
     "AccountDetailsAction",
+    "AccountMentionsAction",
     "PostTweetAction",
+    "PostTweetReplyAction",
     "TWITTER_ACTIONS",
 ]